refactor: Note 모델의 lastModified 타입을 Double로 변경

+++ CloudNotes/CloudNotes/Model/Note.swift
@@ -10,7 +10,7 @@ import Foundation
 struct Note: Decodable {
     let title: String
     let body: String
-    let lastModified: Int
+    let lastModified: Double
     
     enum CodingKeys: String, CodingKey {
         case title, body
